fixed eclipse warnings
git-svn-id: https://svn.apache.org/repos/asf/incubator/pivot/trunk@830099 13f79535-47bb-0310-9956-ffa450edef68

+++ wtk/src/org/apache/pivot/wtk/Theme.java
@@ -171,7 +171,7 @@ public abstract class Theme {
 
                 if (string.endsWith("%")) {
                     float percentage = Float.parseFloat(string.substring(0, string.length() - 1)) / 100f;
-                    size = Math.round((float)font.getSize() * percentage);
+                    size = Math.round(font.getSize() * percentage);
                 } else {
                     throw new IllegalArgumentException(value + " is not a valid font size.");
                 }
